[REFACTOR]db연결 코드 모듈화로 리팩토링
